refactor: View와 Controller 분리
