+++ src/js/model/Model.js
@@ -1,0 +1,1 @@
+export default class Model {}
